Refactor: Introduce CarSnapshot and update ShiftV2 model
- Introduced `CarSnapshot` data class to represent car state during a shift.
- Replaced `Car` with `CarSnapshot` in `ShiftV2`.
- Updated `ShiftV2` to store `carId` instead of the full `Car` object.
- Modified `ShiftMoney` to use `Long` for monetary values (representing cents).
- Added utility functions in `RoundUtils.kt` for converting between cents and dollars, and for rounding doubles.
- Updated `ShiftV2.toShift()` to correctly convert cents to dollars and format mileage.
- Moved `Car.kt` to the `car` subpackage.

+++ app/src/main/java/com/hsact/taxilog/data/model/shift2/ShiftMoney.kt
@@ -1,12 +1,12 @@
 package com.hsact.taxilog.data.model.shift2
 
 data class ShiftMoney(
-    val earnings: Double,
-    val tips: Double = 0.0,
-    val tax: Double = 0.0,
-    val wash: Double = 0.0,
-    val fuelCost: Double,
+    val earnings: Long,
+    val tips: Long = 0,
+    val tax: Long = 0,
+    val wash: Long = 0,
+    val fuelCost: Long,
 ){
-    val profit: Double
+    val profit: Long
         get() = earnings + tips - tax - wash - fuelCost
 }
